test(router): isolate TTS route contracts
