+++ tests/unit_test/router/test_tts_routes.py
@@ -1,0 +1,1299 @@
+from __future__ import annotations
+
+import asyncio
+import json
+import time
+from typing import Any
+
+import httpx
+import pytest
+from fastapi import Request
+from fastapi.testclient import TestClient
+from starlette.datastructures import URL
+
+from sglang_omni_router import proxy as proxy_module
+from sglang_omni_router import websocket_proxy as websocket_proxy_module
+from sglang_omni_router.app import create_app
+from sglang_omni_router.config import RouterConfig, WorkerConfig
+from sglang_omni_router.selector import WorkerSelector
+from sglang_omni_router.voice_routing import VoiceRoutingState
+from sglang_omni_router.worker import Worker, build_workers
+
+
+def _request_netloc(request: httpx.Request) -> str:
+    return f"{request.url.host}:{request.url.port}"
+
+
+def _router_config(
+    *,
+    max_payload_size: int = 512 * 1024 * 1024,
+    health_failure_threshold: int = 1,
+    health_check_interval_secs: int = 10,
+    worker_configs: list[WorkerConfig] | None = None,
+    voice_owner_worker_url: str | None = None,
+) -> RouterConfig:
+    return RouterConfig(
+        workers=worker_configs
+        or [
+            WorkerConfig(url="http://worker-a:8101"),
+            WorkerConfig(url="http://worker-b:8102"),
+        ],
+        max_payload_size=max_payload_size,
+        health_success_threshold=1,
+        health_failure_threshold=health_failure_threshold,
+        health_check_interval_secs=health_check_interval_secs,
+        voice_owner_worker_url=voice_owner_worker_url,
+    )
+
+
+def _voice_routing(
+    config: RouterConfig,
+    workers: list[Worker],
+    client: httpx.AsyncClient,
+) -> VoiceRoutingState:
+    return VoiceRoutingState(
+        workers=workers,
+        owner_url=config.voice_owner_worker_url,
+        client=client,
+        timeout_secs=config.health_check_timeout_secs,
+        retry_interval_secs=config.health_check_interval_secs,
+    )
+
+
+def test_voice_owner_config_must_identify_a_capable_worker() -> None:
+    with pytest.raises(
+        ValueError,
+        match="voice_owner_worker_url must identify a configured worker",
+    ):
+        _router_config(voice_owner_worker_url="http://worker-c:8103")
+
+    with pytest.raises(
+        ValueError,
+        match="must identify a worker with capabilities: audio_input, speech",
+    ):
+        _router_config(
+            worker_configs=[
+                WorkerConfig(
+                    url="http://worker-a:8101",
+                    capabilities={"speech"},
+                )
+            ],
+            voice_owner_worker_url="http://worker-a:8101",
+        )
+
+
+@pytest.mark.asyncio
+async def test_automatic_voice_owner_tracks_dynamic_worker_registration() -> None:
+    config = _router_config(
+        worker_configs=[WorkerConfig(url="http://worker-a:8101", capabilities={"chat"})]
+    )
+    workers = build_workers(config.workers)
+    async with httpx.AsyncClient() as client:
+        state = VoiceRoutingState(
+            workers=workers,
+            owner_url=None,
+            client=client,
+            timeout_secs=5,
+            retry_interval_secs=1,
+        )
+        assert state.resolve_owner() is None
+        assert not state.requires_owner({"preset"})
+
+        workers.append(
+            build_workers(
+                [
+                    WorkerConfig(
+                        url="http://worker-b:8102",
+                        capabilities={"speech", "audio_input"},
+                    )
+                ]
+            )[0]
+        )
+        assert state.resolve_owner() is workers[1]
+        assert state.is_owner(workers[1])
+        assert state.requires_owner({"preset"})
+        workers[0].replace_config(
+            WorkerConfig(
+                url=workers[0].url,
+                capabilities={"speech", "audio_input"},
+            )
+        )
+        assert state.resolve_owner() is workers[1]
+
+
+def test_voice_registry_mutation_before_hydration_preserves_owner_state() -> None:
+    async def run() -> None:
+        def handler(request: httpx.Request) -> httpx.Response:
+            assert request.url.path == "/v1/audio/voices"
+            return httpx.Response(
+                200,
+                json={"uploaded_voices": [{"name": "Existing"}]},
+                request=request,
+            )
+
+        workers = build_workers(_router_config().workers)
+        workers[0].state = "healthy"
+        async with httpx.AsyncClient(transport=httpx.MockTransport(handler)) as client:
+            state = VoiceRoutingState(
+                workers=workers,
+                owner_url=workers[0].url,
+                client=client,
+                timeout_secs=5,
+                retry_interval_secs=1,
+            )
+            state.record_upload("New")
+            await state.start()
+            try:
+                for _ in range(100):
+                    if not state.requires_owner({"preset"}):
+                        break
+                    await asyncio.sleep(0.01)
+                assert not state.requires_owner({"preset"})
+                assert state.requires_owner({"existing"})
+                assert state.requires_owner({"new"})
+            finally:
+                await state.stop()
+
+    asyncio.run(run())
+
+
+def test_voice_registry_mutations_during_hydration_are_not_overwritten() -> None:
+    async def run() -> None:
+        hydration_started = asyncio.Event()
+        release_hydration = asyncio.Event()
+
+        async def handler(request: httpx.Request) -> httpx.Response:
+            assert request.url.path == "/v1/audio/voices"
+            hydration_started.set()
+            await release_hydration.wait()
+            return httpx.Response(
+                200,
+                json={
+                    "uploaded_voices": [
+                        {"name": "Deleted"},
+                        {"name": "Existing"},
+                    ]
+                },
+                request=request,
+            )
+
+        workers = build_workers(_router_config().workers)
+        workers[0].state = "healthy"
+        async with httpx.AsyncClient(transport=httpx.MockTransport(handler)) as client:
+            state = VoiceRoutingState(
+                workers=workers,
+                owner_url=workers[0].url,
+                client=client,
+                timeout_secs=5,
+                retry_interval_secs=1,
+            )
+            await state.start()
+            await hydration_started.wait()
+            state.record_upload("New")
+            state.record_delete("Deleted")
+            release_hydration.set()
+            try:
+                for _ in range(100):
+                    if not state.requires_owner({"preset"}):
+                        break
+                    await asyncio.sleep(0.01)
+                assert not state.requires_owner({"preset"})
+                assert state.requires_owner({"existing"})
+                assert state.requires_owner({"new"})
+                assert not state.requires_owner({"deleted"})
+            finally:
+                await state.stop()
+
+    asyncio.run(run())
+
+
+def test_voice_registry_hydration_retries_without_request_path_io() -> None:
+    async def run() -> None:
+        request_count = 0
+
+        def handler(request: httpx.Request) -> httpx.Response:
+            nonlocal request_count
+            assert request.url.path == "/v1/audio/voices"
+            request_count += 1
+            if request_count == 1:
+                return httpx.Response(503, request=request)
+            return httpx.Response(
+                200,
+                json={"uploaded_voices": []},
+                request=request,
+            )
+
+        workers = build_workers(_router_config().workers)
+        workers[0].state = "healthy"
+        async with httpx.AsyncClient(transport=httpx.MockTransport(handler)) as client:
+            state = VoiceRoutingState(
+                workers=workers,
+                owner_url=workers[0].url,
+                client=client,
+                timeout_secs=5,
+                retry_interval_secs=0.01,
+            )
+            await state.start()
+            try:
+                assert state.requires_owner({"preset"})
+                for _ in range(100):
+                    if request_count >= 2 and not state.requires_owner({"preset"}):
+                        break
+                    await asyncio.sleep(0.01)
+                assert request_count == 2
+                assert not state.requires_owner({"preset"})
+            finally:
+                await state.stop()
+
+    asyncio.run(run())
+
+
+def test_tts_http_routes_preserve_batch_identity_and_uploaded_voice_owner() -> None:
+    seen: list[tuple[str, str, str]] = []
+
+    def handler(request: httpx.Request) -> httpx.Response:
+        worker = _request_netloc(request)
+        path = request.url.path
+        if path == "/health":
+            return httpx.Response(200, json={"status": "healthy"}, request=request)
+        seen.append((request.method, worker, path))
+        if path == "/v1/audio/voices" and request.method == "POST":
+            return httpx.Response(200, json={"name": "Clone"}, request=request)
+        if path == "/v1/audio/voices" and request.method == "GET":
+            return httpx.Response(
+                200,
+                json={"uploaded_voices": [{"name": "Clone"}]},
+                request=request,
+            )
+        if path == "/v1/audio/voices/Clone" and request.method == "DELETE":
+            return httpx.Response(200, json={"success": True}, request=request)
+        if path == "/v1/audio/speech/batch":
+            return httpx.Response(
+                200,
+                json={
+                    "id": "batch-1",
+                    "results": [],
+                    "total": 0,
+                    "succeeded": 0,
+                    "failed": 0,
+                },
+                request=request,
+            )
+        if path == "/v1/audio/speech":
+            status = 400 if worker == "worker-a:8101" else 200
+            return httpx.Response(
+                status,
+                content=b"audio" if status == 200 else b'{"error": {}}',
+                request=request,
+            )
+        raise AssertionError(f"unexpected request: {request.method} {path}")
+
+    async_client = httpx.AsyncClient(transport=httpx.MockTransport(handler))
+    app = create_app(
+        _router_config(
+            voice_owner_worker_url="http://worker-b:8102",
+            health_check_interval_secs=1,
+        ),
+        client=async_client,
+    )
+
+    with TestClient(app) as client:
+        for _ in range(200):
+            if not app.state.voice_routing.requires_owner({"preset"}):
+                break
+            time.sleep(0.01)
+        assert not app.state.voice_routing.requires_owner({"preset"})
+        upload = client.post(
+            "/v1/audio/voices",
+            data={"name": "Clone", "consent": "true"},
+            files={"audio_sample": ("clone.wav", b"RIFF", "audio/wav")},
+        )
+        speech = client.post(
+            "/v1/audio/speech",
+            json={"input": "hello", "voice": "Clone"},
+        )
+        batch = client.post(
+            "/v1/audio/speech/batch",
+            json={
+                "voice": "default",
+                "items": [
+                    {"input": "one"},
+                    {"input": "two", "voice": "Clone"},
+                ],
+            },
+        )
+        deleted = client.delete("/v1/audio/voices/Clone")
+        deleted_voice = client.post(
+            "/v1/audio/speech",
+            json={"input": "hello", "voice": "Clone"},
+        )
+
+    assert upload.status_code == 200
+    assert speech.status_code == 200
+    assert batch.status_code == 200
+    assert deleted.status_code == 200
+    assert deleted_voice.status_code == 400
+    assert seen == [
+        ("GET", "worker-b:8102", "/v1/audio/voices"),
+        ("POST", "worker-b:8102", "/v1/audio/voices"),
+        ("POST", "worker-b:8102", "/v1/audio/speech"),
+        ("POST", "worker-b:8102", "/v1/audio/speech/batch"),
+        ("DELETE", "worker-b:8102", "/v1/audio/voices/Clone"),
+        ("POST", "worker-a:8101", "/v1/audio/speech"),
+    ]
+    owner = app.state.workers[1].to_dict()
+    assert owner["routed_requests_by_class"] == {
+        "voice_control": 2,
+        "speech_http": 1,
+        "speech_batch": 1,
+    }
+
+
+@pytest.mark.asyncio
+async def test_voice_mutations_commit_in_dispatch_order() -> None:
+    upload_started = asyncio.Event()
+    delete_started = asyncio.Event()
+    release_upload = asyncio.Event()
+    dispatched: list[str] = []
+
+    async def handler(request: httpx.Request) -> httpx.Response:
+        if request.method == "GET":
+            return httpx.Response(
+                200,
+                json={"uploaded_voices": []},
+                request=request,
+            )
+        dispatched.append(request.method)
+        if request.method == "POST":
+            upload_started.set()
+            await release_upload.wait()
+        else:
+            delete_started.set()
+        return httpx.Response(200, json={"ok": True}, request=request)
+
+    def request(method: str, path: str) -> Request:
+        boundary = "voice-boundary"
+        body = (
+            (
+                f"--{boundary}\r\n"
+                'Content-Disposition: form-data; name="name"\r\n\r\n'
+                "Clone\r\n"
+                f"--{boundary}--\r\n"
+            ).encode()
+            if method == "POST"
+            else b""
+        )
+        messages = [{"type": "http.request", "body": body, "more_body": False}]
+
+        async def receive() -> dict[str, Any]:
+            if messages:
+                return messages.pop(0)
+            return {"type": "http.request", "body": b"", "more_body": False}
+
+        headers = (
+            [
+                (
+                    b"content-type",
+                    f"multipart/form-data; boundary={boundary}".encode(),
+                )
+            ]
+            if method == "POST"
+            else []
+        )
+        return Request(
+            {
+                "type": "http",
+                "method": method,
+                "path": path,
+                "headers": headers,
+                "query_string": b"",
+                "scheme": "http",
+                "server": ("testserver", 80),
+                "client": ("testclient", 50000),
+            },
+            receive,
+        )
+
+    config = _router_config()
+    workers = build_workers(config.workers)
+    workers[0].state = "healthy"
+    async with httpx.AsyncClient(
+        transport=httpx.MockTransport(handler)
+    ) as async_client:
+        voice_routing = _voice_routing(config, workers, async_client)
+        proxy = proxy_module.ProxyHandler(
+            config=config,
+            workers=workers,
+            selector=WorkerSelector(config.policy),
+            client=async_client,
+            voice_routing=voice_routing,
+        )
+        upload_task = asyncio.create_task(
+            proxy.forward_model_request(
+                request("POST", "/v1/audio/voices"),
+                "/v1/audio/voices",
+            )
+        )
+        await upload_started.wait()
+        delete_task = asyncio.create_task(
+            proxy.forward_model_request(
+                request("DELETE", "/v1/audio/voices/Clone"),
+                "/v1/audio/voices/Clone",
+            )
+        )
+        await asyncio.sleep(0)
+        assert not delete_started.is_set()
+
+        release_upload.set()
+        upload_response, delete_response = await asyncio.gather(
+            upload_task,
+            delete_task,
+        )
+        assert dispatched == ["POST", "DELETE"]
+        for response in (upload_response, delete_response):
+            assert response.status_code == 200
+            assert (
+                b"".join([chunk async for chunk in response.body_iterator])
+                == b'{"ok":true}'
+            )
+
+        await voice_routing.start()
+        try:
+            for _ in range(100):
+                if not voice_routing.requires_owner({"preset"}):
+                    break
+                await asyncio.sleep(0.01)
+            assert not voice_routing.requires_owner({"clone"})
+        finally:
+            await voice_routing.stop()
+
+
+@pytest.mark.parametrize("upload_status", [200, 400])
+@pytest.mark.asyncio
+async def test_voice_upload_mutation_uses_successful_bounded_multipart_request(
+    upload_status: int,
+) -> None:
+    boundary = "quoted-voice-boundary"
+    body = (
+        f"--{boundary}\r\n"
+        'Content-Disposition: form-data; name="audio"; filename="sample.wav"\r\n'
+        "Content-Type: audio/wav\r\n\r\n"
+        "audio-bytes\r\n"
+        f"--{boundary}\r\n"
+        'Content-Disposition: form-data; name="name"\r\n\r\n'
+        "Clone\r\n"
+        f"--{boundary}--\r\n"
+    ).encode()
+    messages = [{"type": "http.request", "body": body, "more_body": False}]
+
+    async def receive() -> dict[str, Any]:
+        if messages:
+            return messages.pop(0)
+        return {"type": "http.request", "body": b"", "more_body": False}
+
+    request = Request(
+        {
+            "type": "http",
+            "method": "POST",
+            "path": "/v1/audio/voices",
+            "headers": [
+                (
+                    b"content-type",
+                    f'multipart/form-data; boundary="{boundary}"'.encode(),
+                )
+            ],
+            "query_string": b"",
+            "scheme": "http",
+            "server": ("testserver", 80),
+            "client": ("testclient", 50000),
+        },
+        receive,
+    )
+
+    def handler(upstream_request: httpx.Request) -> httpx.Response:
+        if upstream_request.method == "GET":
+            return httpx.Response(
+                200,
+                json={"uploaded_voices": []},
+                request=upstream_request,
+            )
+        return httpx.Response(
+            upload_status,
+            json={"name": "Clone"},
+            request=upstream_request,
+        )
+
+    config = _router_config()
+    workers = build_workers(config.workers)
+    workers[0].state = "healthy"
+    async with httpx.AsyncClient(
+        transport=httpx.MockTransport(handler)
+    ) as async_client:
+        voice_routing = _voice_routing(config, workers, async_client)
+        proxy = proxy_module.ProxyHandler(
+            config=config,
+            workers=workers,
+            selector=WorkerSelector(config.policy),
+            client=async_client,
+            voice_routing=voice_routing,
+        )
+        await voice_routing.start()
+        try:
+            for _ in range(100):
+                if not voice_routing.requires_owner({"preset"}):
+                    break
+                await asyncio.sleep(0.01)
+            assert not voice_routing.requires_owner({"clone"})
+
+            response = await proxy.forward_model_request(request, "/v1/audio/voices")
+            assert response.status_code == upload_status
+            assert (
+                b"".join([chunk async for chunk in response.body_iterator])
+                == b'{"name":"Clone"}'
+            )
+            assert voice_routing.requires_owner({"clone"}) is (upload_status == 200)
+        finally:
+            await voice_routing.stop()
+
+
+@pytest.mark.asyncio
+async def test_voice_upload_ownership_is_visible_before_response_body_completes() -> (
+    None
+):
+    release_upload_body = asyncio.Event()
+    boundary = "voice-boundary"
+    body = (
+        f"--{boundary}\r\n"
+        'Content-Disposition: form-data; name="name"\r\n\r\n'
+        "Clone\r\n"
+        f"--{boundary}--\r\n"
+    ).encode()
+    request_messages = [{"type": "http.request", "body": body, "more_body": False}]
+
+    async def receive() -> dict[str, Any]:
+        if request_messages:
+            return request_messages.pop(0)
+        return {"type": "http.request", "body": b"", "more_body": False}
+
+    request = Request(
+        {
+            "type": "http",
+            "method": "POST",
+            "path": "/v1/audio/voices",
+            "headers": [
+                (
+                    b"content-type",
+                    f"multipart/form-data; boundary={boundary}".encode(),
+                )
+            ],
+            "query_string": b"",
+            "scheme": "http",
+            "server": ("testserver", 80),
+            "client": ("testclient", 50000),
+        },
+        receive,
+    )
+
+    class StalledUploadBody(httpx.AsyncByteStream):
+        async def __aiter__(self):
+            await release_upload_body.wait()
+            yield b'{"name":"Clone"}'
+
+    def handler(upstream_request: httpx.Request) -> httpx.Response:
+        if upstream_request.method == "GET":
+            return httpx.Response(
+                200,
+                json={"uploaded_voices": []},
+                request=upstream_request,
+            )
+        return httpx.Response(
+            200,
+            stream=StalledUploadBody(),
+            request=upstream_request,
+        )
+
+    config = _router_config(voice_owner_worker_url="http://worker-b:8102")
+    workers = build_workers(config.workers)
+    workers[1].state = "healthy"
+    async with httpx.AsyncClient(
+        transport=httpx.MockTransport(handler)
+    ) as async_client:
+        voice_routing = _voice_routing(config, workers, async_client)
+        proxy = proxy_module.ProxyHandler(
+            config=config,
+            workers=workers,
+            selector=WorkerSelector(config.policy),
+            client=async_client,
+            voice_routing=voice_routing,
+        )
+        await voice_routing.start()
+        try:
+            for _ in range(100):
+                if not voice_routing.requires_owner({"preset"}):
+                    break
+                await asyncio.sleep(0.01)
+            assert not voice_routing.requires_owner({"preset"})
+
+            response = await proxy.forward_model_request(request, "/v1/audio/voices")
+            assert response.status_code == 200
+            assert voice_routing.requires_owner({"Clone"})
+            release_upload_body.set()
+            assert b"".join([chunk async for chunk in response.body_iterator]) == (
+                b'{"name":"Clone"}'
+            )
+        finally:
+            release_upload_body.set()
+            await voice_routing.stop()
+
+
+class _FakeTTSUpstream:
+    def __init__(self, messages: list[str | bytes]) -> None:
+        self.sent: list[str | bytes] = []
+        self.messages = messages
+        self.close_code = 1000
+        self.close_reason = ""
+
+    async def __aenter__(self):
+        return self
+
+    async def __aexit__(self, exc_type, exc, traceback) -> None:
+        return None
+
+    def __aiter__(self):
+        return self
+
+    async def __anext__(self):
+        if not self.messages:
+            raise StopAsyncIteration
+        return self.messages.pop(0)
+
+    async def send(self, message: str | bytes) -> None:
+        self.sent.append(message)
+
+    async def close(self, *, code: int = 1000) -> None:
+        self.close_code = code
+
+
+@pytest.mark.asyncio
+async def test_tts_websocket_rejects_oversized_followup_before_upstream_send() -> None:
+    class OversizedClient:
+        application_state = websocket_proxy_module.WebSocketState.CONNECTED
+
+        async def receive(self) -> dict[str, Any]:
+            return {"type": "websocket.receive", "text": "x" * 5}
+
+    class WaitingUpstream(_FakeTTSUpstream):
+        async def __anext__(self):
+            await asyncio.Event().wait()
+
+    upstream = WaitingUpstream([])
+    outcome = await websocket_proxy_module._relay(
+        OversizedClient(),
+        upstream,
+        max_client_message_bytes=4,
+    )
+
+    assert outcome == "client_message_too_large"
+    assert upstream.sent == []
+    assert upstream.close_code == 1009
+
+
+@pytest.mark.asyncio
+async def test_tts_websocket_relay_propagates_simultaneous_secondary_failure() -> None:
+    class ConnectedWebSocket:
+        application_state = websocket_proxy_module.WebSocketState.CONNECTED
+
+    class Upstream:
+        close_code = 1000
+        close_reason = ""
+
+    async def client_result():
+        return "client_disconnected"
+
+    async def upstream_failure():
+        raise AssertionError("simultaneous relay defect")
+
+    client_task = asyncio.create_task(client_result())
+    upstream_task = asyncio.create_task(upstream_failure())
+    await asyncio.sleep(0)
+    assert client_task.done()
+    assert upstream_task.done()
+
+    with pytest.raises(AssertionError, match="simultaneous relay defect"):
+        await websocket_proxy_module._coordinate_relay(
+            ConnectedWebSocket(),
+            Upstream(),
+            client_task=client_task,
+            upstream_task=upstream_task,
+        )
+
+
+def test_tts_websocket_is_pinned_and_counted_on_one_worker(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    upstream = _FakeTTSUpstream(
+        [
+            json.dumps({"type": "session.configured"}),
+            b"pcm",
+            json.dumps({"type": "session.done"}),
+        ]
+    )
+
+    def fake_connect(*args, **kwargs):
+        assert args[0] == "ws://worker-a:8101/v1/audio/speech/stream"
+        assert kwargs["compression"] is None
+        assert kwargs["max_queue"] == 1
+        assert kwargs["max_size"] == 512 * 1024 * 1024
+        return upstream
+
+    monkeypatch.setattr(websocket_proxy_module, "websocket_connect", fake_connect)
+
+    def handler(request: httpx.Request) -> httpx.Response:
+        if request.url.path == "/health":
+            return httpx.Response(200, json={"status": "healthy"}, request=request)
+        if request.url.path == "/v1/audio/voices":
+            return httpx.Response(200, json={"uploaded_voices": []}, request=request)
+        raise AssertionError(f"unexpected HTTP request path: {request.url.path}")
+
+    async_client = httpx.AsyncClient(transport=httpx.MockTransport(handler))
+    app = create_app(_router_config(max_payload_size=1024), client=async_client)
+
+    with TestClient(app) as client:
+        with client.websocket_connect("/v1/audio/speech/stream") as websocket:
+            websocket.send_json(
+                {
+                    "type": "session.config",
+                    "model": None,
+                    "voice": "default",
+                }
+            )
+            assert websocket.receive_json()["type"] == "session.configured"
+            assert websocket.receive_bytes() == b"pcm"
+            assert websocket.receive_json()["type"] == "session.done"
+
+    assert len(upstream.sent) == 1
+    assert json.loads(upstream.sent[0])["type"] == "session.config"
+    worker = app.state.workers[0]
+    assert worker.active_requests == 0
+    assert worker.routed_requests_by_class == {"tts_websocket": 1}
+    assert worker.successful_requests_by_class == {"tts_websocket": 1}
+
+
+def test_tts_websocket_protocol_rejection_does_not_evict_worker(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    upstream = _FakeTTSUpstream(
+        [
+            json.dumps(
+                {
+                    "type": "error",
+                    "message": "first WebSocket message must be session.config",
+                    "error_type": "invalid_request_error",
+                    "param": "type",
+                }
+            )
+        ]
+    )
+    monkeypatch.setattr(
+        websocket_proxy_module,
+        "websocket_connect",
+        lambda *args, **kwargs: upstream,
+    )
+
+    def handler(request: httpx.Request) -> httpx.Response:
+        if request.url.path == "/health":
+            return httpx.Response(200, json={"status": "healthy"}, request=request)
+        if request.url.path == "/v1/audio/voices":
+            return httpx.Response(200, json={"uploaded_voices": []}, request=request)
+        raise AssertionError(f"unexpected HTTP request path: {request.url.path}")
+
+    app = create_app(
+        _router_config(health_failure_threshold=1),
+        client=httpx.AsyncClient(transport=httpx.MockTransport(handler)),
+    )
+
+    with TestClient(app) as client:
+        with client.websocket_connect("/v1/audio/speech/stream") as websocket:
+            websocket.send_json({"type": "not.session.config"})
+            error = websocket.receive_json()
+            assert error["type"] == "error"
+            assert error["error_type"] == "invalid_request_error"
+            assert websocket.receive()["type"] == "websocket.close"
+
+    worker = app.state.workers[0]
+    assert worker.is_healthy
+    assert worker.consecutive_failures == 0
+    assert worker.failed_requests_by_class == {"tts_websocket": 1}
+
+
+def test_tts_websocket_sender_close_drains_terminal_success(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    class UpstreamConnectionClosed(Exception):
+        pass
+
+    class ClosingUpstream:
+        close_code = 1000
+        close_reason = ""
+
+        def __init__(self) -> None:
+            self.send_count = 0
+            self.send_observed_close = asyncio.Event()
+            self.messages = [
+                json.dumps(
+                    {
+                        "type": "error",
+                        "message": "recoverable input error",
+                        "error_type": "invalid_request_error",
+                    }
+                ),
+                json.dumps({"type": "session.done"}),
+            ]
+
+        async def __aenter__(self):
+            return self
+
+        async def __aexit__(self, exc_type, exc, traceback) -> None:
+            return None
+
+        def __aiter__(self):
+            return self
+
+        async def __anext__(self):
+            await self.send_observed_close.wait()
+            if not self.messages:
+                raise StopAsyncIteration
+            return self.messages.pop(0)
+
+        async def send(self, message: str | bytes) -> None:
+            self.send_count += 1
+            if self.send_count > 1:
+                self.send_observed_close.set()
+                raise UpstreamConnectionClosed
+
+    upstream = ClosingUpstream()
+    monkeypatch.setattr(
+        websocket_proxy_module,
+        "websocket_connect",
+        lambda *args, **kwargs: upstream,
+    )
+    monkeypatch.setattr(
+        websocket_proxy_module,
+        "ConnectionClosed",
+        UpstreamConnectionClosed,
+    )
+
+    def handler(request: httpx.Request) -> httpx.Response:
+        if request.url.path == "/health":
+            return httpx.Response(200, json={"status": "healthy"}, request=request)
+        if request.url.path == "/v1/audio/voices":
+            return httpx.Response(200, json={"uploaded_voices": []}, request=request)
+        raise AssertionError(f"unexpected HTTP request path: {request.url.path}")
+
+    app = create_app(
+        _router_config(health_failure_threshold=1),
+        client=httpx.AsyncClient(transport=httpx.MockTransport(handler)),
+    )
+
+    with TestClient(app) as client:
+        with client.websocket_connect("/v1/audio/speech/stream") as websocket:
+            websocket.send_json(
+                {
+                    "type": "session.config",
+                    "voice": "default",
+                }
+            )
+            websocket.send_json({"type": "input.text", "text": "late message"})
+            assert websocket.receive_json()["type"] == "error"
+            assert websocket.receive_json()["type"] == "session.done"
+            assert websocket.receive()["type"] == "websocket.close"
+
+    worker = app.state.workers[0]
+    assert worker.is_healthy
+    assert worker.consecutive_failures == 0
+    assert worker.successful_requests_by_class == {"tts_websocket": 1}
+
+
+def test_tts_websocket_failed_audio_is_not_counted_as_success(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    upstream = _FakeTTSUpstream(
+        [
+            json.dumps(
+                {
+                    "type": "error",
+                    "message": "generation failed",
+                    "error_type": "server_error",
+                }
+            ),
+            json.dumps({"type": "audio.done", "error": True}),
+            json.dumps({"type": "session.done"}),
+        ]
+    )
+    monkeypatch.setattr(
+        websocket_proxy_module,
+        "websocket_connect",
+        lambda *args, **kwargs: upstream,
+    )
+
+    def handler(request: httpx.Request) -> httpx.Response:
+        if request.url.path == "/health":
+            return httpx.Response(200, json={"status": "healthy"}, request=request)
+        if request.url.path == "/v1/audio/voices":
+            return httpx.Response(200, json={"uploaded_voices": []}, request=request)
+        raise AssertionError(f"unexpected HTTP request path: {request.url.path}")
+
+    app = create_app(
+        _router_config(health_failure_threshold=1),
+        client=httpx.AsyncClient(transport=httpx.MockTransport(handler)),
+    )
+
+    with TestClient(app) as client:
+        with client.websocket_connect("/v1/audio/speech/stream") as websocket:
+            websocket.send_json({"type": "session.config", "voice": "default"})
+            assert websocket.receive_json()["type"] == "error"
+            assert websocket.receive_json()["type"] == "audio.done"
+            assert websocket.receive_json()["type"] == "session.done"
+            assert websocket.receive()["type"] == "websocket.close"
+
+    worker = app.state.workers[0]
+    assert worker.is_healthy
+    assert worker.consecutive_failures == 0
+    assert worker.failed_requests_by_class == {"tts_websocket": 1}
+
+
+@pytest.mark.asyncio
+async def test_tts_websocket_client_disconnect_is_counted_as_aborted(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    class WaitingUpstream(_FakeTTSUpstream):
+        def __init__(self) -> None:
+            super().__init__([json.dumps({"type": "session.configured"})])
+            self.closed = asyncio.Event()
+
+        async def __anext__(self):
+            if self.messages:
+                return self.messages.pop(0)
+            await self.closed.wait()
+            raise StopAsyncIteration
+
+        async def close(self, *, code: int = 1000) -> None:
+            self.close_code = code
+            self.closed.set()
+
+    class DisconnectingWebSocket:
+        def __init__(self) -> None:
+            self.application_state = websocket_proxy_module.WebSocketState.CONNECTED
+            self.client_state = websocket_proxy_module.WebSocketState.CONNECTED
+            self.headers: dict[str, str] = {}
+            self.url = URL("ws://router/v1/audio/speech/stream")
+            self.first_message = True
+            self.downstream_send_started = asyncio.Event()
+
+        async def accept(self) -> None:
+            return None
+
+        async def receive(self) -> dict[str, Any]:
+            if self.first_message:
+                self.first_message = False
+                return {
+                    "type": "websocket.receive",
+                    "text": json.dumps({"type": "session.config", "voice": "default"}),
+                }
+            await self.downstream_send_started.wait()
+            return {"type": "websocket.disconnect"}
+
+        async def send_text(self, _message: str) -> None:
+            self.downstream_send_started.set()
+            raise OSError("client disconnected during send")
+
+        async def close(self, *, code: int, reason: str = "") -> None:
+            self.application_state = websocket_proxy_module.WebSocketState.DISCONNECTED
+
+    upstream = WaitingUpstream()
+    monkeypatch.setattr(
+        websocket_proxy_module,
+        "websocket_connect",
+        lambda *args, **kwargs: upstream,
+    )
+
+    config = _router_config(health_failure_threshold=1)
+    workers = build_workers(config.workers)
+    workers[0].state = "healthy"
+    async_client = httpx.AsyncClient(transport=httpx.MockTransport(lambda _: None))
+    voice_routing = _voice_routing(config, workers, async_client)
+    proxy = proxy_module.ProxyHandler(
+        config=config,
+        workers=workers,
+        selector=WorkerSelector(config.policy),
+        client=async_client,
+        voice_routing=voice_routing,
+    )
+    websocket_proxy = websocket_proxy_module.TTSWebSocketProxy(
+        config=config,
+        workers=workers,
+        selector=WorkerSelector(config.policy),
+        admission=proxy.admission,
+        voice_routing=voice_routing,
+    )
+    try:
+        await websocket_proxy.forward(DisconnectingWebSocket())
+    finally:
+        await async_client.aclose()
+
+    assert workers[0].is_healthy
+    assert workers[0].consecutive_failures == 0
+    assert workers[0].failed_requests_by_class == {"tts_websocket": 1}
+
+
+@pytest.mark.parametrize(
+    (
+        "handshake_status",
+        "expected_error_type",
+        "expected_close_code",
+        "worker_remains_healthy",
+    ),
+    [
+        (401, "upstream_error", 1008, True),
+        (429, "overloaded_error", 1013, True),
+        (500, "upstream_error", 1011, True),
+        (502, "upstream_error", 1011, False),
+        (503, "server_error", 1013, True),
+    ],
+)
+def test_tts_websocket_handshake_status_controls_worker_health(
+    monkeypatch: pytest.MonkeyPatch,
+    handshake_status: int,
+    expected_error_type: str,
+    expected_close_code: int,
+    worker_remains_healthy: bool,
+) -> None:
+    class RejectedHandshake(websocket_proxy_module.WebSocketException):
+        class Response:
+            def __init__(self, status_code: int) -> None:
+                self.status_code = status_code
+
+        response = Response(handshake_status)
+
+    class RejectedConnection:
+        async def __aenter__(self):
+            raise RejectedHandshake
+
+        async def __aexit__(self, exc_type, exc, traceback) -> None:
+            return None
+
+    monkeypatch.setattr(
+        websocket_proxy_module,
+        "websocket_connect",
+        lambda *args, **kwargs: RejectedConnection(),
+    )
+
+    def handler(request: httpx.Request) -> httpx.Response:
+        if request.url.path == "/health":
+            return httpx.Response(200, json={"status": "healthy"}, request=request)
+        if request.url.path == "/v1/audio/voices":
+            return httpx.Response(200, json={"uploaded_voices": []}, request=request)
+        raise AssertionError(f"unexpected HTTP request path: {request.url.path}")
+
+    app = create_app(
+        _router_config(health_failure_threshold=1),
+        client=httpx.AsyncClient(transport=httpx.MockTransport(handler)),
+    )
+
+    with TestClient(app) as client:
+        with client.websocket_connect("/v1/audio/speech/stream") as websocket:
+            websocket.send_json({"type": "session.config", "voice": "default"})
+            error = websocket.receive_json()
+            close = websocket.receive()
+
+    assert error["type"] == "error"
+    assert error["error_type"] == expected_error_type
+    assert error["code"] == handshake_status
+    assert close["type"] == "websocket.close"
+    assert close["code"] == expected_close_code
+
+    worker = app.state.workers[0]
+    assert worker.is_healthy is worker_remains_healthy
+    assert worker.consecutive_failures == (0 if worker_remains_healthy else 1)
+    assert worker.failed_requests_by_class == {"tts_websocket": 1}
+
+
+@pytest.mark.asyncio
+async def test_tts_websocket_unexpected_relay_failure_propagates(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    class FailingUpstream(_FakeTTSUpstream):
+        async def __anext__(self):
+            raise AssertionError("unexpected relay defect")
+
+    class WaitingWebSocket:
+        application_state = websocket_proxy_module.WebSocketState.CONNECTED
+        client_state = websocket_proxy_module.WebSocketState.CONNECTED
+        headers: dict[str, str] = {}
+        url = URL("ws://router/v1/audio/speech/stream")
+
+        def __init__(self) -> None:
+            self._first_message = True
+
+        async def accept(self) -> None:
+            return None
+
+        async def receive(self) -> dict[str, Any]:
+            if self._first_message:
+                self._first_message = False
+                return {
+                    "type": "websocket.receive",
+                    "text": json.dumps({"type": "session.config", "voice": "default"}),
+                }
+            await asyncio.Event().wait()
+            raise AssertionError("unreachable")
+
+        async def send_text(self, _message: str) -> None:
+            return None
+
+        async def send_bytes(self, _message: bytes) -> None:
+            return None
+
+        async def close(self, *, code: int, reason: str = "") -> None:
+            self.application_state = websocket_proxy_module.WebSocketState.DISCONNECTED
+
+    monkeypatch.setattr(
+        websocket_proxy_module,
+        "websocket_connect",
+        lambda *args, **kwargs: FailingUpstream([]),
+    )
+    config = _router_config(health_failure_threshold=1)
+    workers = build_workers(config.workers)
+    workers[0].state = "healthy"
+    async with httpx.AsyncClient() as async_client:
+        voice_routing = _voice_routing(config, workers, async_client)
+        proxy = proxy_module.ProxyHandler(
+            config=config,
+            workers=workers,
+            selector=WorkerSelector(config.policy),
+            client=async_client,
+            voice_routing=voice_routing,
+        )
+        websocket_proxy = websocket_proxy_module.TTSWebSocketProxy(
+            config=config,
+            workers=workers,
+            selector=WorkerSelector(config.policy),
+            admission=proxy.admission,
+            voice_routing=voice_routing,
+        )
+
+        with pytest.raises(AssertionError, match="unexpected relay defect"):
+            await websocket_proxy.forward(WaitingWebSocket())
+
+    assert workers[0].is_healthy
+    assert workers[0].consecutive_failures == 0
+    assert workers[0].active_requests == 0
+    assert workers[0].routed_requests == 0
+    assert proxy.admission.inflight == 0
+
+
+def test_tts_websocket_policy_close_does_not_evict_worker(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    class PolicyConnectionClosed(Exception):
+        class ReceivedClose:
+            code = 1008
+
+        rcvd = ReceivedClose()
+
+    class PolicyClosingUpstream(_FakeTTSUpstream):
+        def __init__(self) -> None:
+            super().__init__([])
+            self.close_code = 1008
+
+        async def __anext__(self):
+            raise PolicyConnectionClosed
+
+    upstream = PolicyClosingUpstream()
+    monkeypatch.setattr(
+        websocket_proxy_module,
+        "websocket_connect",
+        lambda *args, **kwargs: upstream,
+    )
+    monkeypatch.setattr(
+        websocket_proxy_module,
+        "ConnectionClosed",
+        PolicyConnectionClosed,
+    )
+
+    def handler(request: httpx.Request) -> httpx.Response:
+        if request.url.path == "/health":
+            return httpx.Response(200, json={"status": "healthy"}, request=request)
+        if request.url.path == "/v1/audio/voices":
+            return httpx.Response(200, json={"uploaded_voices": []}, request=request)
+        raise AssertionError(f"unexpected HTTP request path: {request.url.path}")
+
+    app = create_app(
+        _router_config(health_failure_threshold=1),
+        client=httpx.AsyncClient(transport=httpx.MockTransport(handler)),
+    )
+
+    with TestClient(app) as client:
+        with client.websocket_connect("/v1/audio/speech/stream") as websocket:
+            websocket.send_json({"type": "session.config", "voice": "default"})
+            assert websocket.receive()["type"] == "websocket.close"
+
+    worker = app.state.workers[0]
+    assert worker.is_healthy
+    assert worker.consecutive_failures == 0
+    assert worker.failed_requests_by_class == {"tts_websocket": 1}
+
+
+def test_tts_websocket_policy_close_during_initial_send_is_application_failure(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    class PolicyConnectionClosed(websocket_proxy_module.WebSocketException):
+        class ReceivedClose:
+            code = 1008
+
+        rcvd = ReceivedClose()
+
+    class PolicyClosingUpstream(_FakeTTSUpstream):
+        async def send(self, message: str | bytes) -> None:
+            raise PolicyConnectionClosed
+
+    monkeypatch.setattr(
+        websocket_proxy_module,
+        "websocket_connect",
+        lambda *args, **kwargs: PolicyClosingUpstream([]),
+    )
+    monkeypatch.setattr(
+        websocket_proxy_module,
+        "ConnectionClosed",
+        PolicyConnectionClosed,
+    )
+
+    def handler(request: httpx.Request) -> httpx.Response:
+        if request.url.path == "/health":
+            return httpx.Response(200, json={"status": "healthy"}, request=request)
+        if request.url.path == "/v1/audio/voices":
+            return httpx.Response(200, json={"uploaded_voices": []}, request=request)
+        raise AssertionError(f"unexpected HTTP request path: {request.url.path}")
+
+    app = create_app(
+        _router_config(health_failure_threshold=1),
+        client=httpx.AsyncClient(transport=httpx.MockTransport(handler)),
+    )
+
+    with TestClient(app) as client:
+        with client.websocket_connect("/v1/audio/speech/stream") as websocket:
+            websocket.send_json({"type": "session.config", "voice": "default"})
+            error = websocket.receive_json()
+            close = websocket.receive()
+
+    assert error["type"] == "error"
+    assert error["error_type"] == "upstream_error"
+    assert error["code"] == 400
+    assert close["type"] == "websocket.close"
+    assert close["code"] == 1008
+
+    worker = app.state.workers[0]
+    assert worker.is_healthy
+    assert worker.consecutive_failures == 0
+    assert worker.active_requests == 0
+    assert worker.failed_requests_by_class == {"tts_websocket": 1}
